Finish predict function

+++ src/ukf.h
@@ -73,7 +73,9 @@ public:
 
   ///* the current NIS for laser
   double NIS_laser_;
-  long previous_timestamp_;
+  
+  // previous timestamp
+  long long previous_timestamp_;
   /**
    * Constructor
    */
